[Fix] 관리자 제품 등록 - 이미지 등록 오류 수정
