+++ techUpFrontend/components/admin/admin-devmodify.vue
@@ -238,7 +238,6 @@ const submitForm = async () => {
   });
 }
 onMounted(async () => {
-  console.log(idx.value);
   await adminStore.loadProductInfo(idx.value);
 });
 </script>
